+++ cookbooks/wt_streaming_barebones/metadata.rb
@@ -1,0 +1,8 @@
+name                "wt_streaming_barebones"
+maintainer          "Webtrends, Inc"
+maintainer_email    "mark.alexander@webtrends.com"
+license             "All rights reserved"
+description         "Installs/Configures the Webtrends Streaming Bare Bones SAPI utility"
+long_description    IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version             "0.0.3"
+depends             "java"