+++ cookbooks/wt_stream_processor/CHANGELOG.md
@@ -1,3 +1,9 @@
+##1.0.4
+* Classpath changes
+
+##1.0.3
+* Upgrade with Akka defaults
+
 ##1.0.2
 * Fixed runit issues with new runit cookbook