+++ cookbooks/wt_streamingconfigservice/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.2
+* Added connection string for data config endpoints
+
+## 1.1.1
+* Start service as final deploy step.
+
 ## 1.1.0
 * Added support for runit.